moved some files from task1
