+++ src/main/java/ru/faang/school/task_1/Entities/Swordman.java
@@ -1,4 +1,4 @@
-package ru.faang.school.task_1;
+package ru.faang.school.task_1.Entities;
 
 public class Swordman extends Creature{
     public Swordman(){
